+++ cibo/actions/commands/inventory.py
@@ -2,7 +2,7 @@
 
 from typing import List
 
-from cibo.actions._base_ import Action
+from cibo.actions import Action
 from cibo.exceptions import ClientNotLoggedIn
 from cibo.models.client import Client
 from cibo.models.message import Message, MessageRoute