+++ cibo/actions/commands/logout.py
@@ -3,7 +3,7 @@
 from time import sleep
 from typing import List, Tuple
 
-from cibo.actions._base_ import Action
+from cibo.actions import Action
 from cibo.actions.connect import Connect
 from cibo.exceptions import ClientNotLoggedIn
 from cibo.models.client import Client