+++ cibo/actions/commands/move.py
@@ -2,7 +2,7 @@
 
 from typing import List, Tuple
 
-from cibo.actions._base_ import Action
+from cibo.actions import Action
 from cibo.actions.commands.look import Look
 from cibo.exceptions import (
     ClientNotLoggedIn,