+++ cibo/entities/rooms.py
@@ -6,7 +6,7 @@ This is a collection of all the rooms that exist in the world.
 
 from typing import List
 
-from cibo.entities._base_ import Entity
+from cibo.entities import Entity
 from cibo.entities.sectors import Sectors
 from cibo.exceptions import RoomNotFound
 from cibo.models.description import RoomDescription