+++ cibo/events/connect.py
@@ -1,12 +1,14 @@
 """Clients who have connected to the server, since last update poll."""
 
+from typing import Any, Optional
+
 from cibo.actions.connect import Connect
-from cibo.events._base_ import Event
+from cibo.events import Event, EventPayload
 
 
 class ConnectEvent(Event):
     """Clients who have connected to the server, since last update poll."""
 
-    def process(self) -> None:
-        for client in self._telnet.get_new_clients():
-            Connect(self._server_config).process(client, None, [])
+    def process(self, _sender: Any, payload: Optional[EventPayload]) -> None:
+        if payload and payload.client:
+            Connect(self._server_config).process(payload.client, None, [])