+++ cibo/events/disconnect.py
@@ -1,13 +1,15 @@
 """Clients who have disconnected from the server, since last update poll."""
 
 
+from typing import Any, Optional
+
 from cibo.actions.disconnect import Disconnect
-from cibo.events._base_ import Event
+from cibo.events import Event, EventPayload
 
 
 class DisconnectEvent(Event):
     """Clients who have disconnected from the server, since last update poll."""
 
-    def process(self) -> None:
-        for client in self._telnet.get_disconnected_clients():
-            Disconnect(self._server_config).process(client, None, [])
+    def process(self, _sender: Any, payload: Optional[EventPayload]) -> None:
+        if payload and payload.client:
+            Disconnect(self._server_config).process(payload.client, None, [])