+++ cibo/models/data/item.py
@@ -7,7 +7,7 @@ from typing import List, Self
 
 from peewee import AutoField, ForeignKeyField, IntegerField
 
-from cibo.models.data._base_ import Model
+from cibo.models.data import Model
 from cibo.models.data.user import User